Create a succesfully netwrok call
create a network call succesfully

+++ app/src/main/java/com/vcco/weatherapp/model/Wind.kt
@@ -3,8 +3,11 @@ package com.vcco.weatherapp.model
 import com.google.gson.annotations.SerializedName
 import com.vcco.weatherapp.model.utils.ModelConstants as Constants
 
+/**
+ * Current Wind conditions in the location
+ */
 data class Wind(
     val speed: Float,
     @SerializedName(Constants.WIND_DIRECTION_RESPONSE)
-    val direction: Float
+    val direction: Int
 )
